refactor: split config ui routes into interface modules
