+++ src/dicton/config_server.py
@@ -559,391 +559,37 @@ def _background_record():
 def create_app():
     """Create FastAPI application."""
     try:
-        from fastapi import FastAPI, Request
-        from fastapi.responses import HTMLResponse, JSONResponse
         from pydantic import create_model
     except ImportError as e:
         raise ImportError(
             "FastAPI not installed. Install with: pip install dicton[configui]"
         ) from e
 
-    app = FastAPI(title="Dicton Setup")
-
     config_fields = dict.fromkeys(CONFIG_STRING_FIELDS, (str | None, None))
     config_fields.update(dict.fromkeys(CONFIG_BOOL_FIELDS, (bool | None, None)))
     ConfigData = create_model("ConfigData", **config_fields)
-
     AutostartData = create_model("AutostartData", enabled=(bool, ...))
-
-    @app.get("/", response_class=HTMLResponse)
-    async def root():
-        return SETUP_HTML_TEMPLATE
-
-    @app.get("/advanced", response_class=HTMLResponse)
-    async def advanced():
-        return ADVANCED_HTML_TEMPLATE
-
-    @app.get("/api/config")
-    async def api_get_config():
-        return JSONResponse(get_current_config())
-
-    @app.post("/api/config")
-    async def api_save_config(data: ConfigData):
-        try:
-            config_dict = data.model_dump(exclude_none=True)
-            save_config(config_dict)
-            _setup_state["first_test_passed"] = False
-            _setup_state["last_test_text"] = ""
-            return {"status": "ok", "setup": build_setup_status()}
-        except Exception as e:
-            import traceback
-
-            print(f"[ERROR] Save config failed: {e}")
-            traceback.print_exc()
-            return JSONResponse({"status": "error", "message": str(e)}, status_code=500)
-
-    @app.get("/api/setup/status")
-    async def api_setup_status():
-        return JSONResponse(build_setup_status())
-
-    @app.post("/api/setup/save")
-    async def api_setup_save(data: ConfigData):
-        return await api_save_config(data)
-
-    @app.post("/api/setup/autostart")
-    async def api_setup_autostart(data: AutostartData):
-        result = set_autostart(data.enabled)
-        payload = {"status": "ok" if result.get("ok") else "error", **result}
-        payload["setup"] = build_setup_status()
-        status_code = 200 if result.get("ok") else 400
-        return JSONResponse(payload, status_code=status_code)
-
-    @app.post("/api/setup/launch")
-    async def api_setup_launch():
-        result = launch_background()
-        payload = {"status": "ok" if result.get("ok") else "error", **result}
-        payload["setup"] = build_setup_status()
-        status_code = 200 if result.get("ok") else 400
-        return JSONResponse(payload, status_code=status_code)
-
-    @app.get("/api/dictionary")
-    async def api_get_dictionary():
-        return JSONResponse(get_dictionary())
-
-    @app.post("/api/dictionary")
-    async def api_add_similarity_word(data: dict):
-        word = data.get("word", "")
-        if word:
-            add_similarity_word(word)
-            return {"status": "ok"}
-        return JSONResponse({"error": "Missing word"}, status_code=400)
-
-    @app.delete("/api/dictionary")
-    async def api_remove_similarity_word(data: dict):
-        word = data.get("word", "")
-        if word:
-            remove_similarity_word(word)
-            return {"status": "ok"}
-        return JSONResponse({"error": "Missing word"}, status_code=400)
-
-    @app.get("/api/context/profiles")
-    async def api_get_context_profiles():
-        """Get list of available context profiles."""
-        try:
-            from .context_profiles import get_profile_manager
-
-            manager = get_profile_manager()
-            manager.load()
-            return list(manager.list_profiles())
-        except Exception as e:
-            return JSONResponse({"error": str(e)}, status_code=500)
-
-    @app.get("/api/context/current")
-    async def api_get_current_context():
-        """Get current context detection result."""
-        try:
-            from .context_detector import get_context_detector
-            from .context_profiles import get_profile_manager
-
-            detector = get_context_detector()
-            if not detector:
-                return {
-                    "app_name": "N/A",
-                    "window_title": "Context detection not available",
-                    "wm_class": "",
-                    "matched_profile": "default",
-                    "typing_speed": "normal",
-                }
-
-            context = detector.get_context()
-            manager = get_profile_manager()
-            profile = manager.match_context(context)
-
-            return {
-                "app_name": context.app_name if context else "",
-                "window_title": context.window.title if context and context.window else "",
-                "wm_class": context.window.wm_class if context and context.window else "",
-                "matched_profile": profile.name if profile else "default",
-                "typing_speed": profile.typing_speed if profile else "normal",
-            }
-        except Exception as e:
-            return JSONResponse({"error": str(e)}, status_code=500)
-
-    @app.get("/api/context/profiles/{profile_name}")
-    async def api_get_profile(profile_name: str):
-        """Get a specific profile's full details."""
-        try:
-            from .context_profiles import get_profile_manager
-
-            manager = get_profile_manager()
-            manager.load()
-            profile = manager.get_profile(profile_name)
-
-            if not profile:
-                return JSONResponse(
-                    {"error": f"Profile '{profile_name}' not found"}, status_code=404
-                )
-
-            return {
-                "name": profile.name,
-                "match": {
-                    "wm_class": profile.match.wm_class,
-                    "window_title_contains": profile.match.window_title_contains,
-                    "file_extension": profile.match.file_extension,
-                    "widget_role": profile.match.widget_role,
-                    "url_contains": profile.match.url_contains,
-                },
-                "llm_preamble": profile.llm_preamble,
-                "typing_speed": profile.typing_speed,
-                "formatting": profile.formatting,
-                "extends": profile.extends,
-                "priority": profile.priority,
-            }
-        except Exception as e:
-            return JSONResponse({"error": str(e)}, status_code=500)
-
-    @app.put("/api/context/profiles/{profile_name}")
-    async def api_update_profile(profile_name: str, request: Request):
-        """Update or create a profile (saved to user config)."""
-        import json
-
-        try:
-            data = await request.json()
-            user_config_path = get_user_contexts_path()
-
-            # Load existing user config or start fresh
-            if user_config_path.exists():
-                with open(user_config_path) as f:
-                    user_config = json.load(f)
-            else:
-                user_config = {"profiles": {}, "typing_speeds": {}}
-
-            # Update/add the profile
-            user_config["profiles"][profile_name] = {
-                "match": data.get("match", {}),
-                "llm_preamble": data.get("llm_preamble", ""),
-                "typing_speed": data.get("typing_speed", "normal"),
-                "formatting": data.get("formatting", "auto"),
-                "priority": data.get("priority", 0),
-            }
-
-            if data.get("extends"):
-                user_config["profiles"][profile_name]["extends"] = data["extends"]
-
-            # Ensure config directory exists
-            user_config_path.parent.mkdir(parents=True, exist_ok=True)
-
-            # Write to user config
-            with open(user_config_path, "w") as f:
-                json.dump(user_config, f, indent=2)
-
-            # Reload profiles
-            from .context_profiles import get_profile_manager
-
-            manager = get_profile_manager()
-            manager.reload()
-
-            return {"status": "ok", "profile": profile_name}
-        except Exception as e:
-            return JSONResponse({"error": str(e)}, status_code=500)
-
-    @app.delete("/api/context/profiles/{profile_name}")
-    async def api_delete_profile(profile_name: str):
-        """Delete a user profile (cannot delete bundled defaults)."""
-        import json
-
-        try:
-            if profile_name == "default":
-                return JSONResponse({"error": "Cannot delete the default profile"}, status_code=400)
-
-            user_config_path = get_user_contexts_path()
-
-            if not user_config_path.exists():
-                return JSONResponse(
-                    {
-                        "error": f"Profile '{profile_name}' is a bundled default and cannot be deleted"
-                    },
-                    status_code=400,
-                )
-
-            with open(user_config_path) as f:
-                user_config = json.load(f)
-
-            if profile_name not in user_config.get("profiles", {}):
-                return JSONResponse(
-                    {
-                        "error": f"Profile '{profile_name}' is a bundled default and cannot be deleted"
-                    },
-                    status_code=400,
-                )
-
-            # Remove from user config
-            del user_config["profiles"][profile_name]
-
-            with open(user_config_path, "w") as f:
-                json.dump(user_config, f, indent=2)
-
-            # Reload profiles
-            from .context_profiles import get_profile_manager
-
-            manager = get_profile_manager()
-            manager.reload()
-
-            return {"status": "ok", "deleted": profile_name}
-        except Exception as e:
-            return JSONResponse({"error": str(e)}, status_code=500)
-
-    @app.post("/api/test/start")
-    async def api_test_start():
-        """Start recording for latency test."""
-        import threading
-        import time
-
-        from .speech_recognition_engine import SpeechRecognizer
-
-        # Clean up any previous test state
-        if _test_state["recording"]:
-            _test_state["recording"] = False
-            if _test_state["record_thread"]:
-                _test_state["record_thread"].join(timeout=1.0)
-
-        # Initialize recognizer
-        _test_state["recognizer"] = SpeechRecognizer()
-        _test_state["frames"] = []
-        _test_state["audio_data"] = None
-        _test_state["start_time"] = time.time()
-        _test_state["recording"] = True
-
-        # Start background recording thread
-        _test_state["record_thread"] = threading.Thread(target=_background_record)
-        _test_state["record_thread"].start()
-
-        return {"status": "recording"}
-
-    @app.post("/api/test/stop")
-    async def api_test_stop():
-        """Stop recording and run transcription test - mirrors exact production flow."""
-        import time
-
-        import numpy as np
-
-        if not _test_state["recording"]:
-            return JSONResponse({"error": "Not recording"}, status_code=400)
-
-        # Stop recording
-        _test_state["recording"] = False
-        record_end_time = time.time()
-        record_duration_ms = (record_end_time - _test_state["start_time"]) * 1000
-
-        # Wait for recording thread to finish
-        if _test_state["record_thread"]:
-            _test_state["record_thread"].join(timeout=2.0)
-
-        # Convert captured frames to audio array
-        frames = _test_state["frames"]
-        if not frames:
-            _setup_state["first_test_passed"] = False
-            return JSONResponse(
-                {"error": "No audio captured", "setup": build_setup_status()},
-                status_code=400,
-            )
-
-        audio_data = b"".join(frames)
-        audio_array = np.frombuffer(audio_data, dtype=np.int16).astype(np.float32) / 32768.0
-
-        # Get the recognizer used for recording (same instance for transcription)
-        recognizer = _test_state["recognizer"]
-        if not recognizer:
-            _setup_state["first_test_passed"] = False
-            return JSONResponse(
-                {"error": "Recognizer not available", "setup": build_setup_status()},
-                status_code=500,
-            )
-
-        result = {
-            "latency": {
-                "recording": record_duration_ms,
-                "stt": 0,
-                "llm": 0,
-                "total": 0,
-            },
-            "text": "",
-            "stt_provider": "ElevenLabs" if recognizer.use_elevenlabs else "None",
-        }
-
-        try:
-            # === STEP 1: STT Transcription (mirrors production) ===
-            stt_start = time.time()
-            text = recognizer.transcribe(audio_array)
-            result["latency"]["stt"] = (time.time() - stt_start) * 1000
-
-            if not text:
-                result["error"] = "No speech detected"
-                _setup_state["first_test_passed"] = False
-                result["setup"] = build_setup_status()
-                return JSONResponse(result)
-
-            # === STEP 2: LLM Processing (mirrors production _process_text for BASIC mode) ===
-            llm_start = time.time()
-            try:
-                from . import llm_processor
-
-                if config.ENABLE_REFORMULATION and llm_processor.is_available():
-                    processed = llm_processor.reformulate(text)
-                    if processed:
-                        text = processed
-                    result["llm_provider"] = config.LLM_PROVIDER.capitalize()
-                else:
-                    result["llm_provider"] = (
-                        "Disabled" if not config.ENABLE_REFORMULATION else "Not configured"
-                    )
-            except ImportError:
-                result["llm_provider"] = "Not available"
-
-            result["latency"]["llm"] = (time.time() - llm_start) * 1000
-
-            result["text"] = text
-            result["latency"]["total"] = (
-                result["latency"]["recording"] + result["latency"]["stt"] + result["latency"]["llm"]
-            )
-            _setup_state["first_test_passed"] = True
-            _setup_state["last_test_text"] = text
-
-        except Exception as e:
-            result["error"] = str(e)
-            _setup_state["first_test_passed"] = False
-
-        finally:
-            # Cleanup
-            if _test_state["recognizer"]:
-                _test_state["recognizer"].cleanup()
-                _test_state["recognizer"] = None
-            _test_state["frames"] = []
-
-        result["setup"] = build_setup_status()
-        return JSONResponse(result)
-
-    return app
+    from .interfaces.web.app import create_web_app
+
+    return create_web_app(
+        setup_html=SETUP_HTML_TEMPLATE,
+        advanced_html=ADVANCED_HTML_TEMPLATE,
+        config_model=ConfigData,
+        autostart_model=AutostartData,
+        dependencies={
+            "get_current_config": get_current_config,
+            "save_config": save_config,
+            "build_setup_status": build_setup_status,
+            "set_autostart": set_autostart,
+            "launch_background": launch_background,
+            "get_dictionary": get_dictionary,
+            "add_similarity_word": add_similarity_word,
+            "remove_similarity_word": remove_similarity_word,
+            "setup_state": _setup_state,
+            "test_state": _test_state,
+            "background_record": _background_record,
+        },
+    )
 
 
 def find_available_port(start_port: int = 6873, max_attempts: int = 10) -> int:
