+++ src/dicton/interfaces/__init__.py
@@ -1,0 +1,1 @@
+"""Interface entrypoints for Dicton."""
