+++ src/dicton/interfaces/web/__init__.py
@@ -1,0 +1,1 @@
+"""Web interface modules for the configuration UI."""
